+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 
 - Apart of the changes from adding `view_*` async builders, we can have a couple breaking changes to the `view_*` functions:
   - `{Account, Contract, Worker}::view_state` moved `prefix` parameter into builder. i.e.
-  - {}
+  - `view` function changed to be a builder, and no longer take in `args` as a parameter. It instead has been moved to the builder side.
+  - Changed `Worker::view_latest_block` to `Worker::view_block` as the default behavior is equivalent.
 
 ```
 worker.view_state("account_id", Some(prefix)).await?;